move to message fragment
